requirements.txt
amqp==2.6.1
billiard==3.6.4.0
celery==4.3.0
Django==2.0
django-environ==0.4.5
djangorestframework==3.9.2
importlib-metadata==4.0.1
kombu==4.6.11
pytz==2021.1
redis==3.2.1
typing-extensions==3.7.4.3
vine==1.3.0
zipp==3.4.1
